pyproject.toml
[project]
name = "inspect"
version = "0.1.0"
description = "Inspect AI evaluation tasks"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "inspect-ai",
    "openai",
    "python-dotenv",
    "pre-commit",
    "ruff",
    "datasets>=4.4.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tasks"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py
